requirements.txt
msal
requests
pandas
openpyxl
beautifulsoup4
python-dateutil
dateparser>=1.2.0
